iio: potentiometer: ds1803: Remove VLA usage
Remove VLA and replace it with fixed a fixed length array and therefore,
prevent potential stack overflow attacks. Since we know that there are
only two channels for the driver so let's make it explicit. Add a small
comment to explain that there are two channels used to store raw data,
read from ds1803_read_raw() into the specific channels(0 or 1).

Fixed as a part of the discussion to remove all VLAs from the kernel :
https://lkml.org/lkml/2018/3/7/621

Cc: keescook@chromium.org
Signed-off-by: Himanshu Jha <himanshujha199640@gmail.com>